single js file, added AutoFillPromos
